Create Prompts_for_agents.py
This is a starting point for people that want to contribute, but don't know much coding or datascience. If you are a good writer or know of some creative or interesting topics, including them in high quality "scripts" or "prompts" for us to use as a chat history as we start to build the agent datasets can be a very valuable contribution.